Move Renderer & SwapChain types to their dedicated headers
