+++ include/nox/renderer/renderer.h
@@ -2,7 +2,7 @@
 
 #include <nox/export.h>
 #include <nox/non_copyable.h>
-#include <nox/renderer/common_types.h>
+#include <nox/renderer/renderer_types.h>
 
 #include <functional>
 #include <memory>
@@ -24,6 +24,8 @@ struct RendererDescriptor {
 
 class NOX_EXPORT Renderer : public NonCopyable {
   public:
+    virtual ~Renderer();
+
     [[nodiscard]] static std::unique_ptr<Renderer, RendererDeleter> create(const RendererDescriptor &descriptor);
 
     virtual std::shared_ptr<SwapChain> createSwapChain(const SwapChainDescriptor &descriptor, const Window &window) = 0;
