+++ src/renderer/swap_chain.cpp
@@ -1,1 +1,7 @@
 #include <nox/renderer/swap_chain.h>
+
+namespace NOX {
+
+SwapChain::~SwapChain() = default;
+
+} // namespace NOX
